safari-534.53-branch/LayoutTests/platform/win/accessibility/parent-element-expected.txt

This tests that elements return their correct parent element.



PASS optionElement.parentElement().role is "list"
